Add format to parameters and move functionality into default plugin
